bot/requirements.txt
google-generativeai
nextcord
python-dotenv
psycopg2-binary
openai